requirements.txt
Flask==2.3.2
pytube==15.0.0
python-vlc==3.0.18121
python-dotenv==1.0.0
gunicorn==20.1.0